Simplify ZarrBarPlotLayer impl by using BarPlotLayer internally
