+++ bindings-js/demo/src/Demo.jsx
@@ -269,31 +269,14 @@ const DEMOS = {
     plot_params: {
       layers: [
         {
-          layer_type: "ZarrBarLayer",
+          layer_type: "ZarrBarPlotLayer",
           layer_params: {
             layer_id: "layer_1",
             store_name: "wheat",
             bounds: null,
             orientation: "Vertical",
-            stroke_width: 5.0,
-
-            categorical_key: "/year",
-            quantitative_key: "/wheat",
-          }
-        },
-        {
-          layer_type: "AxisLinearLayer",
-          layer_params: {
-            layer_id: "left_axis",
-            position: "Left",
-          }
-        },
-        {
-          layer_type: "AxisBandLayer",
-          layer_params: {
-            layer_id: "bottom_axis",
-            position: "Bottom",
-            domain: ["2000", "2001"]
+            identifier_key: "/year",
+            quantity_key: "/wheat",
           }
         },
       ]
