+++ crates/pluot_zarr/src/layers/mod.rs
@@ -1,6 +1,6 @@
 pub mod zarr_point_layer;
 pub mod zarr_point_3d_layer;
-pub mod zarr_bar_layer;
+pub mod zarr_bar_plot_layer;
 pub mod ome_zarr_utils;
 pub mod ome_zarr_bitmap_layer;
 pub mod ome_zarr_multiscale_layer;
